+++ CHANGELOG.md
@@ -83,6 +83,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1354](https://github.com/regen-network/regen-ledger/pull/1354) Add `UpdateClassFees` msg-based gov proposal
 - [#1391](https://github.com/regen-network/regen-ledger/pull/1391) Add `BasketFees` params query
 - [#1412](https://github.com/regen-network/regen-ledger/pull/1412) Add `EventRemoveAllowedDenom`
+- [#1423](https://github.com/regen-network/regen-ledger/pull/1423) Add `AllowedClassCreators` prams query
+- [#1423](https://github.com/regen-network/regen-ledger/pull/1423) Add `CreditClassFees` prams query
+- [#1423](https://github.com/regen-network/regen-ledger/pull/1423) Add `CreditClassAllowlistEnabled` prams query
+
 
 #### Changed